requirements.txt
torch>=1.9.0
transformers>=4.20.0
Pillow>=8.0.0
sv-ttk>=2.0.0

# Optional: For better performance
# torch-audio  # if needed for some torch installations
# torchvision  # if needed for some torch installations